backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
yt-dlp==2023.11.16
google-generativeai==0.3.0
python-multipart==0.0.6
whisper==1.1.10
pydantic==2.5.1